+++ packages/synapse-react-client/src/components/EntityHeaderTable/EntityHeaderTable.stories.ts
@@ -1,9 +1,9 @@
 import { Meta, StoryObj } from '@storybook/react'
-import { MOCK_REPO_ORIGIN } from '../utils/functions/getEndpoint'
-import { getHandlers } from '../mocks/msw/handlers'
+import { MOCK_REPO_ORIGIN } from '../../utils/functions/getEndpoint'
+import { getHandlers } from '../../mocks/msw/handlers'
 import { EntityHeaderTable } from './EntityHeaderTable'
-import { MOCK_FILE_ENTITY_ID } from '../mocks/entity/mockFileEntity'
-import { MOCK_TABLE_ENTITY_ID } from '../mocks/entity/mockTableEntity'
+import { MOCK_FILE_ENTITY_ID } from '../../mocks/entity/mockFileEntity'
+import { MOCK_TABLE_ENTITY_ID } from '../../mocks/entity/mockTableEntity'
 import { ReferenceList } from '@sage-bionetworks/synapse-types'
 
 const meta = {
